审批创建活动添加数据校验方法

+++ activity-service/src/main/java/com/chaoxing/activity/service/queue/activity/handler/MarketActivityDataPushQueueService.java
@@ -130,9 +130,11 @@ public class MarketActivityDataPushQueueService {
         switch (status) {
             case DELETED:
                 // 删除
-                marketActivityDataPushRecordService.delete(activityDataPushRecord.getId());
-                if (formUserId != null) {
-                    wfwFormApiService.deleteFormRecord(formUserId, formId);
+                if (activityDataPushRecord != null) {
+                    marketActivityDataPushRecordService.delete(activityDataPushRecord.getId());
+                    if (formUserId != null) {
+                        wfwFormApiService.deleteFormRecord(formUserId, formId);
+                    }
                 }
                 break;
             default:
